Finished main layout of drawer
Drawer is now completely laid out, but values are still hardcoded. Will
work to remove hardcoded values in separate branch. Added hiding
functionality to drawer, so that it can be closed by clicking on the
hamburger icon. Need to add animation to hamburger icon, so that it
transforms into an x.